Created Java implementations of accuracy, specificity, sensitivity, and configured bdval to use only Java when doing: restat, cross-validation-consensus-features and  evaluate stats on validation set.
Removed ROCR RMSE calculation. Use the Java version instead. The field is preserved to preserve the file formats, but I would not advise to rely on the values that are calculated. 

I can't seem to run R and connect to it on my laptop. Removing the dependency on R for most stats operations should make configuration easier, and speed up performance evaluation runs.
Added junit tests to check that Java implementations calculate the same thing as ROCR, but I can't run them on this machine, so we need to check the next cruisecontrol results. 

Made restat write maqcii-submission.txt file under the results directory that is being restat. 

Extended the prostate example to work with parameter series (I use this to test locally).




git-svn-id: https://pbtech-vc.med.cornell.edu/public/svn/icb/trunk/icb-commons/icb-stats@20998 60cd2ccc-2b05-0410-9566-da273a940bff